Added/Fixed - Testing area for devmode & avoid "!" to trigger bot
What
=================
Added: a testing area which is the location any test messages will show
up in from now on whilst Heimdall is being worked on, and is separate
from production channels, so as not to clutter them with test data.
Fixed: do not allow a lone "!" to trigger Heimdall.

Issue ID: N/A