Added - `status` command & ArmaManager: `updatePointers` functions
What
=================
Added: `status` command for Arma servers to be added to the list of
servers displayed when the command is used, in addition to the image
showing the server banner.
Added: `updatePointers` command to add new arma server instances that
were started from an unknown source, and remove instances that are no
longer active.

Issue ID: #1