Changed - from `arma-manager.js` to FASTER introduced commandlines
What
=================
Changed the `arma-manager.js` implementation which wasn't working
properly on the server for some reason, to now use the commandline
exports that we get from FASTER. FASTER manages the server profiles and
mod downloads for us, and whenever we start a server, it adds the
commandline used to start the server to the clipboard.

I've also managed to make the index.js code is completely independent of
the module used. This means, that all functions that are used are able
to be shared amongst these different modules without having to change
the index.js code at all. This means I can treat the modules that are
used to manage game servers as "drivers", which do the implementation
specific logic, and exposes certain functionality under certain
keywords:
- start
- stop

Issue ID: #1