Added/Changed/Fixed - Embed func & embed command/action func & arguments
What
=================
Added: New embed function that will format and create the embed object
to be returned in Discord.js messages. This is because it's not just the
`!embed` command that uses the embed functionality. Help and the game
server commands do too.
Changed: `!update` command to be "released" so that we can start using
it to update heimdall on the server.
Changed: `!embed` command to use the new `embed()` function.
Changed: `action()` to now only have 2 parameters and not be limited to
only two words for it's action name.
Fixed: `arguments` object being overwritten in some commands, because I
was overwriting the variable instead of merging the values of the old
and new objects.

Issue ID: #6